backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
google-generativeai
python-dotenv==1.0.0
bson==0.5.10
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
pytest==7.4.4
pytest-flask==1.3.0